pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "determinism-audit"
version = "0.1.0"
description = "CLI tool to measure LLM output determinism across providers and configurations"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12.0",
    "litellm>=1.40.0,<2.0.0",
    "pydantic>=2.7.0,<3.0.0",
    "tenacity>=8.3.0",
    "rich>=13.7.0",
    "jinja2>=3.1.0",
    "sentence-transformers>=3.0.0",
    "torch>=2.2.0",
    "python-dotenv>=1.2.2",
]

[project.scripts]
determinism-audit = "determinism_audit.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/determinism_audit"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-recording>=0.13.0",
    "vcrpy>=6.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
src = ["src"]
line-length = 99
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "PIE", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
files = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"
